fix warnings in decode_10

+++ decode/decode_10.c
@@ -9,8 +9,8 @@ int decode_10(DECODE_FUNC_PARAMS) {
     uint8_t p1 = state->rom[++state->pc];
     uint8_t p2 = state->rom[++state->pc];
 
-    uint8_t reg = ((p1 & 0b00000011) << 3) | ((p2 & 0b11100000) >> 5);
-    uint8_t val = p2 & 0b00011111;
+    uint8_t reg = ((p1 & (uint8_t)0b00000011) << 3) | ((p2 & (uint8_t)0b11100000) >> 5);
+    uint8_t val = p2 & (uint8_t)0b00011111;
 
     uint8_t subcode = (p1 & (uint8_t)0b1111100) >> 2;
